requirements.txt
alembic==1.13.2
atomicwrites==1.4.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
chardet==4.0.0
click==8.1.7
colorama==0.4.6
dataclasses==0.6
DateTime==5.5
Deprecated==1.2.14
deprecation==2.1.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gotrue==0.2.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.1
pluggy==1.5.0
postgrest-py==0.4.0
psycopg2==2.9.9
py==1.11.0
Pygments==2.18.0
PyJWT==2.8.0
pypi==2.1
pytest==6.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
realtime-py==0.1.3
redis==5.0.7
requests==2.25.1
rfc3986==1.5.0
rich==13.7.1
sendgrid==6.11.0
sendgrid-python==0.1.1
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
starkbank-ecdsa==2.2.0
supabase-py==0.0.2
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.19
websockets==9.1
Werkzeug==3.0.3
wrapt==1.16.0
WTForms==3.1.2
zope.interface==6.4.post2
